pyproject.toml
[project]
name = "corelink-ai"
version = "0.1.0"
description = "A finance-first A2A reasoning engine for AgentBeats benchmarking"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "uvicorn>=0.38.0",
    "langgraph>=0.5.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.5.0",
    "mcp>=1.26.0",
    "openpyxl>=3.1.5",
    "pypdf>=6.7.5",
    "python-docx>=1.2.0",
    "pdfplumber>=0.11.4",
    "pandas>=2.0.0",
    "yfinance>=0.2.40",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

